+++ src/Telegram.Bot/Types/Message.cs
@@ -374,7 +374,7 @@ public class Message
     /// Optional. Service message: a user was shared with the bot
     /// </summary>
     [JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore)]
-    public UserShared? UserShared { get; set; }
+    public UsersShared? UsersShared { get; set; }
 
     /// <summary>
     /// Optional. Service message: a chat was shared with the bot